refactor(polykybd): extract the OS shortcut-hint table, with an exhaustive equivalence test
keycode_to_disp_overlay() was 310 lines of poly_keymap.c's 4160 — the largest
self-contained block in the file — and it reached into two globals for its inputs.
It moves to hints/os_hints.c as a pure function:

    const uint32_t* os_hint_for_keycode(uint16_t keycode,
                                        uint8_t mods_raw,
                                        uint8_t active_os_packed);

poly_keymap.c keeps a four-line wrapper that supplies poly_layer_t.mods and
poly_sync_t.active_os, so no call site changes. Three side effects of making the
inputs explicit:

  - the POLY_OS_AUTO_FLAG mask moves inside the table, where a caller cannot
    forget it (it was at the single call site before);
  - the led_t parameter is gone — the body never read it in 310 lines;
  - the table now includes quantum_keycodes.h/modifiers.h rather than
    QMK_KEYBOARD_H, so it links in the standalone test harness.

enum poly_os moves to its own poly_os.h (state.h includes it). state.h pulls in
quantum.h and mru.h, which would otherwise make any consumer untestable.

Extracting into a separate translation unit is legitimately NOT codegen-neutral
(inlining, stack slots, compiler switch-table numbering), so compare_codegen.py
cannot be the proof here. Instead hints/tests/os_hints_reference.c carries the
pre-extraction table verbatim from origin/PolyKybd, and the test compares the two
implementations over EVERY (keycode x mods x os) triple — 1.4M comparisons
including all 32 mod-tap modifier encodings — asserting equal content.

make test:polykybd_os_hints — 8 tests. Mutation-tested twice: breaking the
left/right modifier collapse and dropping the POLY_OS_VALUE_MASK each fail the
exhaustive test plus the specific property test for that rule.

Writing the tests turned up two things, both left as-is because this change is
behaviour-preserving:

  - poly_layer_t.mods is QMK's 8-bit PACKED modifier form, but MOD_RCTL & co. in
    enum mods_5bit are a different encoding (0x11, where bit4 is an L/R flag).
    The table only ever uses MOD_L* constants, which are identical in both, so it
    is correct — but a future edit reaching for MOD_RALT here would compare
    against a value that can never occur. Named in the test file.
  - the mod-tap hint block's eleven combined-modifier branches are UNREACHABLE:
    they test (mods & MOD_MASK_CS) == MOD_MASK_CS where mods is the 5-bit mod-tap
    encoding (max 0x1F) and MOD_MASK_CS is 0x33. Only the three single-modifier
    tests and the final else can fire, so a Ctrl+Shift mod-tap shows "¤" and never
    "¤⇧". Pre-existing; pinned by a test that will fail when it is fixed.

split72/split42/revision2 all build; qmk lint --strict passes on both keyboards.
