+++ builddefs/build_test.mk
@@ -69,6 +69,7 @@ include $(QUANTUM_PATH)/os_detection/tests/rules.mk
 include $(QUANTUM_PATH)/sequencer/tests/rules.mk
 include $(QUANTUM_PATH)/wear_leveling/tests/rules.mk
 include modules/polykybd/polymod_ltr559/tests/rules.mk
+include keyboards/polykybd/hints/tests/rules.mk
 include $(QUANTUM_PATH)/logging/print.mk
 include $(PLATFORM_PATH)/test/rules.mk
 ifneq ($(filter $(FULL_TESTS),$(TEST)),)
