+++ keyboards/polykybd/hints/tests/testlist.mk
@@ -1,0 +1,1 @@
+TEST_LIST += polykybd_os_hints
